pushing changes

+++ main.cpp
@@ -6,7 +6,6 @@
 using namespace std;
 struct Priority {
     bool operator()(Patient const & p1, Patient const & p2) {
-        // return "true" if "p1" is ordered before "p2", for example:
         return p1.getPriority() > p2.getPriority();
     }
 };
@@ -14,7 +13,7 @@ int main() {
     vector<Patient> patients;
     Utilities::read(&patients);
     std::priority_queue<Patient, std::vector<Patient>, Priority> pq;
-    for (int i = 0; i < patients.size(); ++i) {
+    for (unsigned long i = 0; i < patients.size(); ++i) {
         pq.push(patients.at(i));
     }
     while (!pq.empty()) {
